FSM-419 : Added fstp vehicle entry, exit time fields validation

+++ fsm/src/main/java/org/egov/fsm/web/model/vehicle/trip/VehicleTrip.java
@@ -136,7 +136,7 @@ public class VehicleTrip   {
   private Long fstpEntryTime = null;
   
   @JsonProperty("fstpExitTime")
-  private Double fstpExitTime = null;
+  private Long fstpExitTime = null;
 
 
 }
